refactor: remove custom rules feature, keep config overrides only
Custom rules (--custom-rules) allowed adding new analysis rules via JSON.
This is removed in favor of maintaining a complete built-in rule set.

Config overrides (--config) for adjusting scores/severity/thresholds are kept.

- Delete src/core/rules/custom/ (loader, schema, index)
- Delete examples/custom-rules.json
- Move config-loader.ts out of custom/ to rules/
- Remove --custom-rules from CLI, MCP, docs
- Update CLAUDE.md, README.md, REFERENCE.md
