--- examples/custom-rules.json
@@ -1,36 +1,0 @@
-[
-  {
-    "id": "icon-not-component",
-    "category": "component",
-    "severity": "blocking",
-    "score": -10,
-    "match": {
-      "type": ["FRAME", "GROUP"],
-      "maxWidth": 48,
-      "maxHeight": 48,
-      "hasChildren": true,
-      "nameContains": "icon"
-    },
-    "message": "\"{name}\" is an icon but not a component — should be componentized",
-    "why": "Icons that are not components cannot be reused consistently across the design system.",
-    "impact": "Developers will hardcode icon SVGs instead of using a shared component.",
-    "fix": "Convert this icon to a component and publish it to the design system library."
-  },
-  {
-    "id": "large-frame-no-auto-layout",
-    "category": "layout",
-    "severity": "risk",
-    "score": -5,
-    "match": {
-      "type": ["FRAME"],
-      "minWidth": 200,
-      "minHeight": 200,
-      "minChildren": 3,
-      "hasAutoLayout": false
-    },
-    "message": "\"{name}\" is a large frame with {type} children but no Auto Layout",
-    "why": "Large frames with multiple children should use Auto Layout for maintainability.",
-    "impact": "Layout changes require manual repositioning of every child element.",
-    "fix": "Apply Auto Layout with appropriate direction and spacing."
-  }
-]
