+++ src/cli/index.ts
@@ -108,10 +108,7 @@ cli.help((sections) => {
     },
     {
       title: "\nCustomization",
-      body: [
-        `  --config <path>         Override rule settings (see: canicode docs config)`,
-        `  --custom-rules <path>   Add custom rules (see: canicode docs rules)`,
-      ].join("\n"),
+      body: `  --config <path>         Override rule settings (see: canicode docs config)`,
     },
     {
       title: "\nExamples",
