+++ src/core/rules/config-loader.test.ts
