--- src/core/rules/custom/index.ts
@@ -1,3 +1,0 @@
-export { loadCustomRules } from "./custom-rule-loader.js";
-export { loadConfigFile, mergeConfigs, type ConfigFile } from "./config-loader.js";
-export { CustomRuleSchema, CustomRulesFileSchema, type CustomRule } from "./custom-rule-schema.js";
